pyproject.toml
[tool.poetry]
name = "bd-sed"
version = "0.1.0"
description = "API de Banco de dados de alunos"
authors = ["Andersonmathema <andersondrade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.26.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
